split types/methods for one dimensional ranges
